refactor(data-loader): replace parseCsv with new one
