--- packages/data-loader/src/parsers/parseCsv2/__tests__/fixtures/withSubtable/expected.json
@@ -1,234 +1,0 @@
-[
-  {
-    "updatedTime": {
-      "value": "2021-02-16T02:43:00Z"
-    },
-    "dropDown": {
-      "value": "sample1"
-    },
-    "creator": {
-      "value": {
-        "code": "username"
-      }
-    },
-    "modifier": {
-      "value": {
-        "code": "username"
-      }
-    },
-    "richText": {
-      "value": "<div><div>rich text editor<br /></div></div><div>rich text editor<br /></div><div>rich text editor<br /></div>"
-    },
-    "singleLineText": {
-      "value": "\"single line text\""
-    },
-    "number": {
-      "value": "8"
-    },
-    "radioButton": {
-      "value": "sample2"
-    },
-    "multiLineText": {
-      "value": "multi\nline\ntext"
-    },
-    "createdTime": {
-      "value": "2021-02-10T06:14:00Z"
-    },
-    "checkBox": {
-      "value": ["\"sample2\""]
-    },
-    "multiSelect": {
-      "value": ["\"sample3\"", "sample4"]
-    },
-    "subTable": {
-      "value": [
-        {
-          "id": "537306",
-          "value": {
-            "subTableText": {
-              "value": "text_line1"
-            },
-            "subTableCheckbox": {
-              "value": ["st_sample1"]
-            },
-            "userSelect": {
-              "value": [
-                {
-                  "code": "sato"
-                },
-                {
-                  "code": "tanaka"
-                }
-              ]
-            },
-            "groupSelect": {
-              "value": [
-                {
-                  "code": "Administrators"
-                }
-              ]
-            },
-            "organizationSelect": {
-              "value": [
-                {
-                  "code": "Development Div"
-                }
-              ]
-            }
-          }
-        },
-        {
-          "id": "537307",
-          "value": {
-            "subTableText": {
-              "value": "text_line2"
-            },
-            "subTableCheckbox": {
-              "value": ["st_sample2"]
-            },
-            "userSelect": {
-              "value": [
-                {
-                  "code": "sato"
-                },
-                {
-                  "code": "tanaka"
-                }
-              ]
-            },
-            "groupSelect": {
-              "value": [
-                {
-                  "code": "Administrators"
-                }
-              ]
-            },
-            "organizationSelect": {
-              "value": [
-                {
-                  "code": "Development Div"
-                }
-              ]
-            }
-          }
-        }
-      ]
-    }
-  },
-  {
-    "updatedTime": {
-      "value": "2021-02-16T02:43:00Z"
-    },
-    "dropDown": {
-      "value": "sample1"
-    },
-    "creator": {
-      "value": {
-        "code": "username"
-      }
-    },
-    "modifier": {
-      "value": {
-        "code": "username"
-      }
-    },
-    "richText": {
-      "value": "<div><div>rich text editor<br /></div></div><div>rich text editor<br /></div><div>rich text editor<br /></div>"
-    },
-    "singleLineText": {
-      "value": "\"single line text\""
-    },
-    "number": {
-      "value": "8"
-    },
-    "radioButton": {
-      "value": "sample2"
-    },
-    "multiLineText": {
-      "value": "multi\nline\ntext"
-    },
-    "createdTime": {
-      "value": "2021-02-10T06:14:00Z"
-    },
-    "checkBox": {
-      "value": ["\"sample2\""]
-    },
-    "multiSelect": {
-      "value": ["\"sample3\"", "sample4"]
-    },
-    "subTable": {
-      "value": [
-        {
-          "id": "537308",
-          "value": {
-            "subTableText": {
-              "value": "text_line1"
-            },
-            "subTableCheckbox": {
-              "value": ["st_sample1"]
-            },
-            "userSelect": {
-              "value": [
-                {
-                  "code": "sato"
-                },
-                {
-                  "code": "tanaka"
-                }
-              ]
-            },
-            "groupSelect": {
-              "value": [
-                {
-                  "code": "Administrators"
-                }
-              ]
-            },
-            "organizationSelect": {
-              "value": [
-                {
-                  "code": "Development Div"
-                }
-              ]
-            }
-          }
-        },
-        {
-          "id": "537309",
-          "value": {
-            "subTableText": {
-              "value": "text_line2"
-            },
-            "subTableCheckbox": {
-              "value": ["st_sample2"]
-            },
-            "userSelect": {
-              "value": [
-                {
-                  "code": "sato"
-                },
-                {
-                  "code": "tanaka"
-                }
-              ]
-            },
-            "groupSelect": {
-              "value": [
-                {
-                  "code": "Administrators"
-                }
-              ]
-            },
-            "organizationSelect": {
-              "value": [
-                {
-                  "code": "Development Div"
-                }
-              ]
-            }
-          }
-        }
-      ]
-    }
-  }
-]
